msm: iommu: [GCC6] Integer not enough to avoid -Wshift-overflow

+++ arch/arm/mach-msm/include/mach/iommu_hw-v1.h
@@ -56,7 +56,7 @@
 #define SET_FIELD(addr, mask, shift, v) \
 do { \
 	int t = readl_relaxed(addr); \
-	writel_relaxed((t & ~((mask) << (shift))) + (((v) & \
+	writel_relaxed((t & ~((unsigned long)(mask) << (shift))) + (((v) & \
 			(mask)) << (shift)), addr); \
 } while (0)
